Added another Test to C19
Added two additional Articles and an additional Author to the test, in order to test the behavior of the program in the following situation:

One author publishes an article citing another article. At this point this citation is a foreign one. Later, the same author publishes another article together with the author of the previously cited article. Now they are coauthors, the earlier citation is no foreign one anymore.