Added Before method to every test
I experienced some weird side-effects when testing. A failing test
sometimes caused all subsequent tests to fail. After long research, I
found out that this behavior is caused by the fact that a failing test
causes an AssertionException not in it’s test method, but when the
program is executed. This causes the After method to be interrupted,
which leads to the terminal not being reset.